Add Notification Utilities And Refactor Duplicates - via Codex
